Remove duplicated D-Bus talk permission
Drop --talk-name=io.github.softfever.OrcaSlicer.InstanceCheck.* as it is
granted by default (see:
https://docs.flatpak.org/en/latest/sandbox-permissions.html#d-bus-access)
and will be blocked in the linter.

Since the discussion during the submission, we realised that an own-name
access by default implies a talk access, so subnames are also granted by
default as well. See the Session Bus Policy section in:
https://docs.flatpak.org/en/latest/flatpak-command-reference.html

> own
>    The application can own the bus name or names (as well as all the above). Available since 0.2.

and also confirmed here independently in :
flatpak/flatpak#5582 (comment)